refactor(parser): extract parser accuracy failure packets (#0000)
